CachyOS Kernel Manager Turkish localisation (#8)
* CachyOS Kernel Manager Turkish localisation

Any feedback or criticism is valued.

* Finalised cachyos-kernel-manager_tr.ts

Minor tweaks like rewording sentences and adjusting a few agent and locatum verbs for proper and common nouns to which they're subject to, wherever they're allowed, in order for them to sound more natural in the target language, without unnecessary pointers and symbols like asterisks and hyphens.

* Minor update: cachyos-kernel-manager_tr.ts

One of the translated sections in my draft was missing from the pushed iteration for some reason. And the letter '-n' was missing from "<translation>RC…" as well.