👷 scx: use scx_loader to start/stop scheds
Take into consideration the scx_loader configuration:
Load the configuration or use default values if it is not present.
Following this, the user has the option to manually edit the
configuration file using their preferred text editor.

If the scx.service is currently running or enabled, disable it to avoid
any potential conflicts with scx_loader.